Make guardrail tests independent of ripgrep binary
