+++ src/__tests__/forbidden-patterns.test.ts
@@ -1,24 +1,9 @@
-import { execFileSync } from 'node:child_process';
 import { describe, expect, it } from 'vitest';
-
-function rg(pattern: string, scope: readonly string[]): string[] {
-  try {
-    const out = execFileSync(
-      'rg',
-      ['-n', '--no-heading', '--color', 'never', '--glob', '*.ts', '--glob', '*.tsx', pattern, ...scope],
-      { encoding: 'utf8', cwd: process.cwd() },
-    ).trim();
-    return out ? out.split('\n').filter(Boolean) : [];
-  } catch (error) {
-    const status = (error as { status?: number }).status;
-    if (status === 1) return [];
-    throw error;
-  }
-}
+import { rgLines } from '../testing/rg-search';
 
 describe('forbidden patterns (v3 hard rules)', () => {
   it('forbids legacy shape2d record writes in frame executors', () => {
-    const matches = rg('writeShape2D\\s*\\(', [
+    const matches = rgLines('writeShape2D\\s*\\(', [
       'src/runtime/ScheduleExecutor.ts',
       'src/runtime/executeFrameStepped.ts',
     ]);
@@ -29,7 +14,7 @@ describe('forbidden patterns (v3 hard rules)', () => {
   });
 
   it('forbids shape2d object unpack helper in render hot path', () => {
-    const matches = rg('readShape2D\\s*\\(', ['src/runtime/RenderAssembler.ts']);
+    const matches = rgLines('readShape2D\\s*\\(', ['src/runtime/RenderAssembler.ts']);
 
     // [LAW:verifiable-goals] Static gate prevents per-instance object unpack
     // churn from reappearing in render grouping loops.
@@ -37,17 +22,17 @@ describe('forbidden patterns (v3 hard rules)', () => {
   });
 
   it('forbids Path2D allocation in runtime/compiler hot modules', () => {
-    const matches = rg('new\\s+Path2D\\s*\\(', ['src/runtime', 'src/compiler']);
+    const matches = rgLines('new\\s+Path2D\\s*\\(', ['src/runtime', 'src/compiler']);
     expect(matches).toEqual([]);
   });
 
   it('forbids f64 storage-class references in compiler/runtime contracts', () => {
-    const matches = rg("storage:\\s*'f64'", ['src/compiler', 'src/runtime']);
+    const matches = rgLines("storage:\\s*'f64'", ['src/compiler', 'src/runtime']);
     expect(matches).toEqual([]);
   });
 
   it('forbids CPU coordinate-bounds scans in the animation hot path', () => {
-    const matches = rg('calculateContentBounds\\s*\\(', ['src/services/AnimationLoop.ts']);
+    const matches = rgLines('calculateContentBounds\\s*\\(', ['src/services/AnimationLoop.ts']);
 
     // [LAW:dataflow-not-control-flow] Frame orchestration must execute the same
     // GPU-first steps each frame; CPU geometry scans are forbidden in-loop.
@@ -55,7 +40,7 @@ describe('forbidden patterns (v3 hard rules)', () => {
   });
 
   it('forbids mode-flag branching in compute orchestration executors', () => {
-    const matches = rg('mode\\s*===', [
+    const matches = rgLines('mode\\s*===', [
       'src/runtime/ScheduleExecutor.ts',
       'src/runtime/executeFrameStepped.ts',
     ]);
@@ -66,7 +51,7 @@ describe('forbidden patterns (v3 hard rules)', () => {
   });
 
   it('forbids GPU full-buffer copy commands in the WebGPU hot path', () => {
-    const matches = rg('copyBufferToBuffer\\s*\\(', ['src/render/webgpu/WebGPURenderer.ts']);
+    const matches = rgLines('copyBufferToBuffer\\s*\\(', ['src/render/webgpu/WebGPURenderer.ts']);
 
     // [LAW:dataflow-not-control-flow] Frame execution keeps one deterministic
     // compute->draw flow without whole-buffer copy detours.
@@ -74,7 +59,7 @@ describe('forbidden patterns (v3 hard rules)', () => {
   });
 
   it('forbids runtime arena reassignment in frame executors', () => {
-    const matches = rg('state\\.arena\\s*=', [
+    const matches = rgLines('state\\.arena\\s*=', [
       'src/runtime/ScheduleExecutor.ts',
       'src/runtime/executeFrameStepped.ts',
     ]);
@@ -85,7 +70,7 @@ describe('forbidden patterns (v3 hard rules)', () => {
   });
 
   it('forbids shape-bank mutation helpers in renderer/assembler hot paths', () => {
-    const matches = rg('writeShapeBank(Header|HandleMetadata)\\s*\\(', [
+    const matches = rgLines('writeShapeBank(Header|HandleMetadata)\\s*\\(', [
       'src/runtime/RenderAssembler.ts',
       'src/render/webgpu/WebGPURenderer.ts',
     ]);
@@ -93,7 +78,7 @@ describe('forbidden patterns (v3 hard rules)', () => {
   });
 
   it('forbids CPU-side direct indirect-args writes in WebGPU renderer', () => {
-    const matches = rg('writeBuffer\\s*\\(\\s*this\\.indirectArgsBuffer', [
+    const matches = rgLines('writeBuffer\\s*\\(\\s*this\\.indirectArgsBuffer', [
       'src/render/webgpu/WebGPURenderer.ts',
     ]);
     expect(matches).toEqual([]);
